Poprawa deserialzatora plików CSV
